data name field changes, encryption option implemented for Room.
Currently we are not handling the change of the password. This should be handled in the future since we currently don't pass the db name as a parameter too. Current behavior is in case of a password change, data doesn't get inserted to the database and the stack trace is logged. p.s. Error message can be improved.

Signed-off-by: Sercan Turkmen <sercanturkmen@outlook.com>